+++ src/Request/SegmentsGetRequest.php
@@ -1,0 +1,47 @@
+<?php
+declare(strict_types=1);
+
+namespace Citilink\ExpertSenderApi\Request;
+
+use Citilink\ExpertSenderApi\Enum\HttpMethod;
+use Citilink\ExpertSenderApi\RequestInterface;
+
+/**
+ * Segments GET Request
+ *
+ * @author Nikita Sapogov <sapogov.n@citilink.ru>
+ */
+class SegmentsGetRequest implements RequestInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function toXml(): string
+    {
+        return '';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getQueryParams(): array
+    {
+        return [];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getMethod(): HttpMethod
+    {
+        return HttpMethod::GET();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getUri(): string
+    {
+        return '/Api/Segments';
+    }
+}